+++ app/javascript/vue/components/radials/object/components/annotatorExtend.js
@@ -28,11 +28,10 @@ const vueAnnotator = {
       loadOnMounted: true
     }
   },
-  mounted: function () {
-    let that = this
-    if(this.loadOnMounted) {
+  created: function () {
+    if (this.loadOnMounted) {
       this.getList((typeof this.urlList == 'undefined') ? `${this.url}/${this.type}.json` : this.urlList).then(response => {
-        that.list = response.body
+        this.list = response.body
       })
     }
   },