+++ app/javascript/vue/components/radials/object/components/collecting_event/ceSection.vue
@@ -1,0 +1,45 @@
+<template>
+  <div>
+    <h3
+      v-if="!!collectingEvent"
+      class="middle">
+      <span v-html="collectingEvent.object_tag"/>
+      <button
+        type="button"
+        class="button circle-button btn-undo button-default"
+        @click="setCE({})"/>
+    </h3>
+    <smart-selector
+      v-else
+      class="full_width"
+      ref="smartSelector"
+      model="collecting_events"
+      target="CollectionObject"
+      klass="CollectionObject"
+      pin-section="CollectingEvents"
+      pin-type="CollectingEvent"
+      @selected="setCE"/>
+  </div>
+</template>
+
+<script>
+
+import SmartSelector from 'components/smartSelector'
+
+export default {
+  components: {
+    SmartSelector
+  },
+  props: {
+    collectingEvent: {
+      type: Object,
+      default: undefined
+    }
+  },
+  methods: {
+    setCE (ce) {
+      this.$emit('onSelect', ce)
+    }
+  }
+}
+</script>