+++ app/javascript/vue/tasks/collection_objects/filter/components/filter.vue
@@ -34,20 +34,39 @@
         Search
       </button>
       <geographic-component
+        class="margin-large-bottom margin-medium-top"
         v-model="params.geographic"/>
-      <otu-component v-model="params.determination"/>
+      <otu-component
+        class="margin-large-bottom"
+        v-model="params.determination"/>
       <collecting-event
+        class="margin-large-bottom"
         v-model="params.collectingEvents"/>
-      <user-component 
+      <user-component
+        class="margin-large-bottom"
         @onUserslist="usersList = $event"
         v-model="params.user"/>
-      <keywords-component v-model="params.keywords.keyword_ids" />
-      <repository-component v-model="params.repository.repository_id"/>
-      <identifier-component v-model="params.identifier"/>
-      <types-component v-model="params.types"/>
-      <loan-component v-model="params.loans"/>
-      <in-relationship v-model="params.relationships.biological_relationship_ids"/>
-      <biocurations-component v-model="params.biocurations.biocuration_class_ids"/>
+      <keywords-component
+        class="margin-large-bottom"
+        v-model="params.keywords.keyword_ids" />
+      <repository-component
+        class="margin-large-bottom"
+        v-model="params.repository.repository_id"/>
+      <identifier-component
+        class="margin-large-bottom"
+        v-model="params.identifier"/>
+      <types-component
+        class="margin-large-bottom"
+        v-model="params.types"/>
+      <loan-component
+        class="margin-large-bottom"
+        v-model="params.loans"/>
+      <in-relationship
+        class="margin-large-bottom"
+        v-model="params.relationships.biological_relationship_ids"/>
+      <biocurations-component
+        class="margin-large-bottom"
+        v-model="params.biocurations.biocuration_class_ids"/>
     </div>
   </div>
 </template>