+++ app/javascript/vue/tasks/nomenclature/new_taxon_name/app.vue
@@ -26,7 +26,7 @@
     <div>
       <nav-header :menu="menu"/>
       <div class="flexbox horizontal-center-content align-start">
-        <div class="ccenter item separate-right">
+        <div class="ccenter item">
           <spinner
             :full-screen="true"
             :legend="(loading ? 'Loading...' : 'Saving changes...')"
@@ -42,7 +42,7 @@
         </div>
         <div
           v-if="getTaxon.id"
-          class="cright item separate-left">
+          class="cright item margin-medium-left">
           <div id="cright-panel">
             <div class="panel content margin-medium-bottom">
               <autocomplete
@@ -241,12 +241,6 @@ export default {
     margin: 0 auto;
     margin-top: 1em;
     max-width: 1240px;
-
-    .basic-information {
-      .vue-autocomplete-input {
-        width: 300px;
-      }
-    }
 
     .autocomplete-search-bar {
       input {