+++ app/javascript/vue/tasks/otu/browse/lang/en.js
@@ -3,6 +3,9 @@
 
 export default {
   section: {
+    header: {
+      validButton: 'Redirect to OTU linked with valid taxon name'
+    },
     status: 'Shows the development status of this section.<br>Green: ready.<br>Yellow: stable<br>Red: prototype',
     options: {
       drag: 'Drag and drop to changes sections order',